+++ packages/bridge-controller/src/index.ts
@@ -138,7 +138,11 @@ export { toQuoteResponseV2 } from './coercers/quote-response-v1-to-v2.js';
 export { toQuoteMetadataV1 } from './utils/quote-metadata/to-quote-metadata-v1.js';
 export { toQuoteMetadataV2 } from './utils/quote-metadata/to-quote-metadata-v2.js';
 
-export { sumAmounts } from './utils/number-formatters.js';
+export {
+  sumAmounts,
+  calcAtomicTokenAmount,
+  calcNormalizedTokenAmount,
+} from './utils/number-formatters.js';
 export { assetIdsMatch } from './utils/assets.js';
 
 export {